+++ MarkdownHelper/scripts/saveAsMDHCleanMindmap.groovy
@@ -1,0 +1,3 @@
+import edofro.MarkDownHelper.MdhCleaner as mc
+
+mc.saveAsMdhCleanMap(node.map)